Add Core.Validator
